+++ lib/helper/routes.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_twitter_clone/page/Auth/selectAuthMethod.dart';
 import 'package:flutter_twitter_clone/page/Auth/verifyEmail.dart';
 import 'package:flutter_twitter_clone/page/common/splash.dart';
+import 'package:flutter_twitter_clone/page/feed/composeTweet/createFeed.dart';
+import 'package:flutter_twitter_clone/page/feed/composeTweet/feedPostreply.dart';
 import 'package:flutter_twitter_clone/page/message/conversationInformation/conversationInformation.dart';
 import 'package:flutter_twitter_clone/page/message/newMessagePage.dart';
 import 'package:flutter_twitter_clone/page/profile/follow/followerListPage.dart';
@@ -21,12 +23,10 @@ import 'package:flutter_twitter_clone/page/settings/accountSettings/proxy/proxyP
 import 'package:flutter_twitter_clone/page/settings/settingsAndPrivacyPage.dart';
 import '../page/Auth/signin.dart';
 import '../helper/customRoute.dart';
-import '../page/feed/createFeed.dart';
 import '../page/feed/imageViewPage.dart';
 import '../page/Auth/forgetPasswordPage.dart';
 import '../page/Auth/signup.dart';
 import '../page/feed/feedPostDetail.dart';
-import '../page/feed/feedPostreply.dart';
 import '../page/profile/EditProfilePage.dart';
 import '../page/message/chatScreenPage.dart';
 import '../page/profile/profilePage.dart';
