+++ lib/page/feed/composeTweet/createFeed.dart
@@ -1,8 +1,5 @@
 import 'dart:io';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_twitter_clone/helper/theme.dart' as prefix0;
-import 'package:flutter_twitter_clone/helper/theme.dart';
 import 'package:flutter_twitter_clone/model/feedModel.dart';
 import 'package:flutter_twitter_clone/helper/constant.dart';
 import 'package:flutter_twitter_clone/helper/utility.dart';
@@ -11,11 +8,9 @@ import 'package:flutter_twitter_clone/state/authState.dart';
 import 'package:flutter_twitter_clone/state/feedState.dart';
 import 'package:flutter_twitter_clone/widgets/customAppBar.dart';
 import 'package:flutter_twitter_clone/widgets/customWidgets.dart';
-import 'package:flutter_twitter_clone/widgets/newWidget/customLoader.dart';
+import 'package:flutter_twitter_clone/widgets/tweet/widgets/bottomIconWidget.dart';
+import 'package:flutter_twitter_clone/widgets/tweet/widgets/tweetImage.dart';
 import 'package:provider/provider.dart';
-
-import 'widgets/bottomIconWidget.dart';
-import 'widgets/tweetImage.dart';
 
 class CreateFeedPage extends StatefulWidget {
   CreateFeedPage({Key key}) : super(key: key);
